first fit, need refinement

+++ src/app/model/TruckLoadingProblem.java
@@ -3,15 +3,15 @@ package app.model;
 import java.util.List;
 
 public class TruckLoadingProblem {
-	private final int binCapacity;
+	private final double binCapacity;
 	private final List<Parcel> parcels;
 	
-	public TruckLoadingProblem(int binCapacity, List<Parcel> parcels) {
+	public TruckLoadingProblem(double binCapacity, List<Parcel> parcels) {
 		this.binCapacity = binCapacity;
 		this.parcels = parcels;
 	}
 	
-	public int getBinCapacity() {
+	public double getBinCapacity() {
 		return binCapacity;
 	}
 	public List<Parcel> getParcels() {
